refactor(frontend): 상태 배지/포맷 함수 공용화
- StatusBadge 컴포넌트 분리: src/app/documents/components/StatusBadge.tsx
- 포맷 유틸 추가: src/lib/ui/format.ts (formatDate/formatSize)
- /documents 및 /documents/[id]에서 공용 모듈 사용
